fix: rename power to powerlaw
