+++ witch/powerlaw.py
